+++ src/components/hero/Hero.styled.tsx
@@ -16,25 +16,3 @@ export const Section = styled.section`
     }
   }
 `;
-
-export const FlexColumn1 = styled.section.attrs<{ $gap?: string }>(() => ({}))`
-  display: flex;
-  flex-direction: column;
-  gap: ${(props) => props.$gap || '1rem'};
-`;
-
-export const FlexColumn = styled.section<{ $gap?: string }>`
-  display: flex;
-  flex-direction: column;
-  gap: ${(props) => props.$gap || '1rem'};
-`;
-
-export const FlexRow = styled.section<{
-  $alignSelf?: 'flex-start' | 'flex-end';
-  $gap?: string;
-}>`
-  align-self: ${(props) => (props.$alignSelf ? props.$alignSelf : 'flex-start')};
-  display: flex;
-  flex-direction: row;
-  gap: ${(props) => props.$gap || '1rem'};
-`;
